requirements.txt
python-decouple==3.8

google-api-python-client==2.165.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.32.3